nilanikan-backend/requirements.txt
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
python-dotenv==1.0.1
Pillow==10.4.0
whitenoise==6.6.0
